separate game figuration to different module
