+++ test/tic_tac_toe_test.exs
@@ -2,40 +2,22 @@ defmodule TictacToeTest do
   use ExUnit.Case
   doctest TicTacToe
 
-  describe ".get_board_size" do
-    test "when the user picks option 1, the game board size is 3x3" do
-      defmodule PicksOptionOne do
-        def get_board_size, do: "1"
-        def print(_message), do: nil
-      end
-      assert TicTacToe.get_board_size(PicksOptionOne) == 3
-    end
-
-    test "when the user picks option 2, the game board size is 4x4" do
-      defmodule PicksOptionTwo do
-        def get_board_size, do: "2"
-        def print(_message), do: nil
-      end
-      assert TicTacToe.get_board_size(PicksOptionTwo) == 4
-    end
-  end
-
   describe ".restart_game" do
-    test "when the input is 'no'" do
+    test "when the user enters 'no'" do
       defmodule PicksOptionNo do
         def print(_message), do: %Message{}.good_bye
       end
       assert TicTacToe.restart_game("no", PicksOptionNo) == %Message{}.good_bye
     end
 
-    test "when the input is 'n'" do
+    test "when the user enters 'n'" do
       defmodule PicksOptionN do
         def print(_message), do: %Message{}.good_bye
       end
       assert TicTacToe.restart_game("n", PicksOptionN) == %Message{}.good_bye
     end
 
-    test "when the input is 'yes'" do
+    test "when the user enters 'yes'" do
       defmodule PicksOptionYes do
         def print(_message), do: nil
       end
@@ -48,7 +30,7 @@ defmodule TictacToeTest do
       assert TicTacToe.restart_game("yes", PicksOptionYes, FakeLoop_1) == "restarting game"
     end
 
-    test "when the input is 'y'" do
+    test "when the user enters 'y'" do
       defmodule PicksOptionY do
         def print(_message), do: nil
       end
@@ -73,56 +55,6 @@ defmodule TictacToeTest do
                   current_player: %Player{mark: "X"},
                   size: 3}
       assert TicTacToe.get_user_move(game, PickMove).board() == [0, 1, 2, 3, "X", 5, 6, 7, 8]
-    end
-  end
-
-  describe ".get_player_marks" do
-    test "when user enters 'W' as mark" do
-      defmodule PickMarkW do
-        def get_player_marks, do: "W"
-        def print(_message), do: nil
-      end
-      assert TicTacToe.get_player_marks(PickMarkW) == "W"
-    end
-
-    test "when user enters 'X' as mark" do
-      defmodule PickMarkX do
-        def get_player_marks, do: "X"
-        def print(_message), do: nil
-      end
-      assert TicTacToe.get_player_marks(PickMarkX) == "X"
-    end
-  end
-
-  describe ".get_marks_for_human_vs_computer_game" do
-    test "when user enters 'X' as mark" do
-      defmodule ComputerPickO do
-        def get_player_marks, do: "X"
-        def print(_message), do: nil
-      end
-      assert TicTacToe.get_marks_for_human_vs_computer_game(ComputerPickO) == {%Player{mark: "X"}, %Computer{mark: "O"}}
-    end
-    test "when user enters 'x' as mark" do
-      defmodule ComputerPickO_2 do
-        def get_player_marks, do: "x"
-        def print(_message), do: nil
-      end
-      assert TicTacToe.get_marks_for_human_vs_computer_game(ComputerPickO_2) == {%Player{mark: "x"}, %Computer{mark: "o"}}
-    end
-  end
-
-  describe ".get_marks_for_human_vs_human_game" do
-    test "when user enters the same marks" do
-      defmodule HumanPick do
-        def get_player_marks, do: "X"
-        def print(_message), do: nil
-      end
-      defmodule Human2_Pick do
-        def get_player_marks, do: "W"
-        def print(_message), do: nil
-      end
-
-      assert TicTacToe.get_marks_for_human_vs_human_game(HumanPick, Human2_Pick) == {%Player{mark: "X"}, %Player{mark: "W"}}
     end
   end
 end
